mininet

+++ cs144_lab3/router/sr_message.c
@@ -32,7 +32,7 @@ void send_arp_packet(struct sr_instance* sr, uint8_t* packet, unsigned int len, 
 	free(_packet);
 }/* end send_arp_packet */
 
-void send_ip_packet(struct sr_instance* sr, uint8_t* packet, char* interface, uint8_t type, uint8_t code){
+void send_ip_packet(struct sr_instance* sr, uint8_t* packet, char* interface, uint16_t type, uint16_t code){
 	int length = sizeof(sr_ethernet_hdr_t) + sizeof(sr_ip_hdr_t) + sizeof(sr_icmp_t3_hdr_t);
 	struct sr_if *interfaces = sr_get_interface(sr, interface);
 
@@ -115,7 +115,7 @@ void build_ip_header(uint8_t *_packet, sr_ip_hdr_t* ip_header, struct sr_if* int
 	ip_tmp_header->ip_sum = cksum(ip_header, sizeof(sr_ip_hdr_t));
 }
 
-void build_icmp_header(uint8_t *_packet, uint8_t *packet, sr_ip_hdr_t *ip_header, sr_icmp_hdr_t* icmp_orig_header, struct sr_if* interfaces, uint8_t type, uint8_t code){
+void build_icmp_header(uint8_t *_packet, uint8_t *packet, sr_ip_hdr_t *ip_header, sr_icmp_hdr_t* icmp_orig_header, struct sr_if* interfaces, uint16_t type, uint16_t code){
 	sr_icmp_t3_hdr_t *icmp_tmp_hdr = (sr_icmp_t3_hdr_t *)(_packet + sizeof(sr_ethernet_hdr_t) + sizeof(sr_ip_hdr_t));
 	icmp_tmp_hdr->icmp_type = type;
 	icmp_tmp_hdr->icmp_code = code;
